Add Write and Flush methods to huffman.Writer
Together, Write() and Flush() allow a huffman.Writer to encode a data
stream using Huffman coding. Note that (at least as implemented now) any
writes to a huffman.Writer via Write() must be followed up with Flush().
This ensures that any remaining bits in currByte (effectively a buffer
of bits) are also written.