Change how 'code' is represented in Codebook
At the moment, 'code' is stored as a binary string. This was originally
done to simplify writing code with leading 0s (e.g. "000"). However,
storing code as a string makes reading much clunkier - bits would need
to be read into a byte buffer, converted to a string, and then compared
to the code.

As an integer only two checks need to be made:
1. Number of bits read == code length
2. Value of byte containing bits == code