Update Writer to handle 'code' as an integer
I've updated Writer to handle the previous change to Codebook. The
update - mostly changing how the code would be written as a series of
bits - was fortunately pretty simple.

Rather than loop over a binary string as before, the Write function now
creates an array of 'bits' equal to the code's length. This array is
filled by repeated getting the least significant bit of the code and
right shifting it by 1.