Add function to derive codebook from header
Once a Reader has finished extracting the header from the encoded file,
DeriveCanonicalCB() is called to create a Codebook using the information
in the header. This Codebook is then used to process the rest of the
file.

NOTE: I've also taken the time to clean up some of the existing code in
codebook.go and codebook_test.go.