Implement Read method for huffman.Reader
A huffman.Reader uses its Read() method to do the following:
1. Read a byte via its underlying bufio.Reader
2. Read through the byte as a series of bits.
   Bits are stored in a buffer (readBuf). When the value in the
   buffer matches a key in huffman.Reader's decoding table (and the
   number of bits read matches the corresponding code length), the
   decoded value given by the table is inserted into the byte array
   being read into.
3. After a successful match, the buffer and number of bits read are
   reset and process continues.

It should be noted that the method is not quite complete. Most
glaringly, it doesn't have a way to terminate correctly after all the
bytes have been read. This will likely require either including the
number of encoded values in the header (to compare against as values are
read into the byte array), or will require creating some king of EOF bit
pattern to check against.
Given the potential size of this change, it will have to wait to be
part of a project-wide refactoring which I'll be starting soon.