Add function to create weight maps
The MakeWeightMap() function creates a map between bytes in a data
stream and their occurrences in said stream. This map can then be used
to create Huffman trees, codebooks, etc.

NOTE: I've added a test case in 'reader_test.go' which includes this
function. This will only be temporary and was done to verify that
reading / writing work with large UTF8 symbols.