Begin experiment with new Header scheme
Currently, the project treats 'Codebooks' and 'Headers' as separate
entities:
1. Codebooks are ordered arrays wherein each element describes a unit,
its code, and that code's length.
2. Headers are the result of extracting data from the Codebook to be put
at the top of encoded files to help with decoding.

After adding huffman.Reader and huffman.Writer, I found that this was
becoming pretty unwieldy. It had resulted in related code spanning
multiple files, and each change to the Codebook or Header type meant
having to rework multiple related functions.

To avoid all of this, I've decided to try to combine the two into a
single Header type. In doing this, all types related to the Header
(including Codebook), as well as all functions, will be organized into 1
file. In addition, I hope to be able to replace all current Header /
Codebook functions with just:
1. NewHeader - Get Header struct from info about units and their weights
2. DeriveHeader - Get Header struct from first n lines of encrypted
   file.
3. Header.String() - Returns a string representation of the Header to be
written during encryption.

It's my hope that by consolidating this functionality, the existing code
will become cleaner, less coupled, and easier to alter in the future.