Add String method to Header
The String() method returns a formatted string describing the Header.
This string will be written to a file during encoding, and be used
during decoding to decode the body of the file.

For the moment, the string consists of four parts:
1. A space separated list of integers describing the number of codes
whose code length == (index in list + 1).
2. A set of all units in the file.
3. The number of units in the file.
4. A delimiter indicating the end of the header (headerDelim).

The idea behind the final delimiter is to allow huffman.Reader's Read()
function to easily extract the Header description. As the structure of
the string generated by String() might change in the future, this
helps to decouple Read() from the method's implementation.