Add function to derive Header from desc string
The DeriveHeader() function parses the header description written to
encoded files and uses that parsed information to recreate the Header
struct. Once this has happened, Header.ExtractDecoder() can be used to
decode the rest of the file.

Currently, reading / parsing the header description and creating a
header from it are two separate events occurring in different files.
Again, this is a level of coupling I'd like to avoid. As such, both
parsing and header creation occur in DeriveHeader(). The huffman.Reader
is now only responsible for reading the header (start of the file to
first instance of 'headerDelim').