Change headerDelim from 2 bytes to one
The current headerDelim is "\r\n". However, as none of the standard
Readers can search for 2 byte patterns, recognizing when this delimiter
has been read is difficult.

To avoid this issue, I've replace headerDelim with just '\n'. As each
line of the header ends with '\n', the end of the Header can be found by
continuously reading to the next '\n' until the amount read equals 1
(i.e. we read a blank line).