Refactor huffman.Reader to work with new Header
The huffman.Reader type and associated functions now use the new Header
type. Much like with huffman.Writer before it, this has resulted in
cleaner, simpler code.

However, there is a problem I've just noticed with this new, refactored
code - it does not appear to handle multi-byte UTF8 characters well.
This is something I'll need to explore in a new branch.