Add function to supplant existing MakeWeightMap
The new ProcessData() function (name subject to change) is nearly
identical to MakeWeightMap(). The only difference is that the former, in
addition to returning a map, also returns an integer given the number of
units (right now bytes) read from the file.

This additional value is used in Headers to indicate when decoding
should stop - i.e. after n units have been decoded.