Update readHeader to reflect header delim change
The current readHeader() function expects the header description string
to be formatted as follows:
1. Lines ending in headerDelim ('\n') which comprise header.
2. Final headerDelim indicating end of header.
To get the entire header, the function repeatedly calls
ReadString(headerDelim) until a string is found containing only the
headerDelim.

While the function appeared to be working, I'd overlooked the fact that
it fails for any input containing '\n'. When this happens, the second
line of the header (which is supposed to be the set of units in the
original file as a single-line string) is separated into two lines.

To address this, I've updated headerDelim to the multi-byte pattern
"\r\n", and have altered readHeader accordingly.