Change headerDelim to handle input with '\n'
Currently, Header cannot handle input containing '\n'. This is because:
1. Each line of the Header description is delimited by '\n'
2. The second line of the Header - which contains all bytes appearing in
   the input - will contain '\n'.
So any attempts to split the Header description using '\n' as a
delimiter will split the second line of the Header.

To avoid this problem, I've changed headerDelim to "\r\n" and
standardized its use within the various Header functions. This solution
should work for most files on a Linux system, and lays the foundation
for setting headerDelim conditionally in the future.