Fix bug in Reader when making multiple calls
When Read() has filled a buffer, the method immediately returns.
If Read() is only called once this isn't a problem, as it satisfies the
requirements of the Reader interface.
However, when Read() needs to be called multiple times, it can lose bits
between calls. This happens when the buffer is filled before Read()
finishes looping over all the bits in the current read byte.

To fix this, I've added some properties to Reader to retain these final
bits between calls. When Read() is next called, these bits will be used
first before another new byte is read.

I've also made the following small change to Reader:
1. Return on io.EOF now correctly returns 'n' as number of bytes read
   instead of 0.