Improve Write() by using internal buffer
Currently, Write() writes bytes individually as they are created. This
is very slow. To improve Write()'s performance, I've added an internal
buffer to store encoded bytes as they are written. Once the buffer is
filled, the entire buffer is written to Writer's underlying io.Writer.
This cuts the number of writes significantly and thus boosts
performance.

It should be noted that this buffer exists on a per-call basis.
Everything stored within the buffer is guaranteed to be written before
the function completes.