Fix bug in NewHeader() for empty input
Currently, NewHeader() assigns canonical Huffman codes by performing a
loop, assigning codes to indices up to len(codebook) - 2, and then
performs a final assignment to index len(codebook) - 1.

If the unitWeights argument to NewHeader() is empty (either because
NewHeader() was called with empty arg or NewHeaderFromReader() is
reading from empty source), then the second assignment above will result
in an error. This is because len(codebook) is 0 in such cases.

To avoid this, I've surrounded the second assignment with an
if-statement to ensure that the index being used is less than
len(codebook).